requirements.txt
asgiref==3.8.1
Django==5.1.4
django-environ==0.11.2
Faker==33.3.1
pillow==11.1.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
qrcode==7.4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
python-telegram-bot==20.7
